Fix the Create User page
The page that is used to create a new user account previously
failed to accept POST requests. This was because the cookie-parser
package hadn't been installed and configured. As a result, the
CSRF cookie couldn't be retrieved. This has been rectified.

Furthermore, a JavaScript file was added to the page that prompts
users for their mobile phone number. This file sends an AJAX
request to the server that creates a 6-digit one-time-password
to be sent to the user via text message. The request handler for
this AJAX endpoint hasn't been written yet.