Write API endpoint to send SMS code
The route handler was written that sends a one-time password to
a user's phone to verify their phone number. Furthermore, there
were a few problems with the way Passport was configured. This
resulted in the user's information not being stored properly
in their session data. These problems have been rectified and
the user's session information is stored properly between
requests.